requirements.txt
apache-airflow==2.10.2
pandas
requests
beautifulsoup4
